dhis2_web_service/flask/requirements.txt
blinker==1.6.3
click==8.1.7
Flask
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug
uwsgi==2.0.22
Flask-HTTPAuth==4.8.0
pandas
psycopg2-binary
